♻️[승훈]refactor: 컴포넌트 형식 통일
