+++ src/App.jsx
@@ -1,8 +1,8 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Main from './pages/Main';
-import List from './pages/List/List';
+import List from './pages/List';
 import Paper from './pages/Paper';
-import Message from './pages/Message/Message';
+import Message from './pages/Message';
 import NavLayout from './layout/NavLayout';
 import Post from './pages/PostPage';
