+++ src/components/TextField/Dropdown/index.js
@@ -1,1 +1,1 @@
-export * from './Dropdown';
+export { default } from './Dropdown';
