+++ src/components/TextField/Input/index.js
@@ -1,1 +1,1 @@
-export * from './Input';
+export * from './InputStyled';
