+++ src/components/TextField/TextEditor/index.js
@@ -1,1 +1,1 @@
-export * from './TextEditor';
+export { default } from './TextEditor';
