+++ src/pages/List/components/Card/index.js
@@ -1,0 +1,1 @@
+export { default } from './Card';
