+++ src/pages/List/components/CardContent/index.js
@@ -1,0 +1,1 @@
+export { default } from './CardContent';
