+++ src/pages/Message/component/EnterContent/index.js
@@ -1,0 +1,1 @@
+export { default } from './EnterContent';
