+++ src/pages/Message/component/FontSelect/index.js
@@ -1,0 +1,1 @@
+export { default } from './FontSelect';
