+++ src/pages/Message/component/PostInput/index.js
@@ -1,0 +1,1 @@
+export { default } from './PostInput';
