+++ src/pages/Message/component/ProfileImageSelect/index.js
@@ -1,0 +1,1 @@
+export { default } from './ProfileImageSelect';
