+++ src/pages/Message/component/RelationshipSelect/RelationshipSelect.jsx
@@ -1,8 +1,7 @@
 import { useState, useEffect } from 'react';
-import TextFields from '../../../components/TextField/TextFields';
-import styled from 'styled-components';
-import COLORS from '../../../utils/colors';
-import FONTS from '../../../utils/Fonts';
+import TextFields from '../../../../components/TextField/TextFields';
+import * as S from './RelationshipSelectStyled';
+
 const OPTIONS = [
   { value: '지인', name: '지인' },
   { value: '동료', name: '동료' },
@@ -10,7 +9,7 @@ const OPTIONS = [
   { value: '친구', name: '친구' },
 ];
 
-export const RelationshipSelect = ({ messageData, setMessageData }) => {
+const RelationshipSelect = ({ messageData, setMessageData }) => {
   const [isError, setIsError] = useState(false);
 
   const handleChange = (selectedOption) => {
@@ -35,14 +34,9 @@ export const RelationshipSelect = ({ messageData, setMessageData }) => {
         onChange={handleChange}
         isError={isError}
       />
-      <ErrorMessage isError={isError}>상대와의 관계를 선택해주세요.</ErrorMessage>
+      <S.ErrorMessage isError={isError}>상대와의 관계를 선택해주세요.</S.ErrorMessage>
     </>
   );
 };
 
-const ErrorMessage = styled.p`
-  ${FONTS.font12_Regular}
-  color: ${COLORS.error};
-  visibility: ${({ isError }) => (isError ? 'visible' : 'hidden')};
-  margin-top: 0.4rem;
-`;
+export default RelationshipSelect;
