+++ src/pages/Message/component/RelationshipSelect/index.js
@@ -1,0 +1,1 @@
+export { default } from './RelationshipSelect';
