--- src/pages/Message/component/index.js
@@ -1,5 +1,0 @@
-export { PostInput } from './PostInput';
-export { RelationshipSelect } from './RelationshipSelect';
-export { EnterContent } from './EnterContent';
-export { FontSelect } from './FontSelect';
-export { ProfileImageSelect } from './ProfileImageSelect';
