+++ src/pages/Message/index.js
@@ -1,0 +1,1 @@
+export { default } from './Message';
